requirements.txt
# ── Web Framework ─────────────────────────────────────────────────────────────
Flask==3.1.0
Flask-Login==0.6.3
Flask-WTF==1.2.2
WTForms==3.2.2
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-Limiter==3.9.4
Flask-Caching==2.3.1
Flask-CORS==5.0.1
gunicorn==23.0.0

# ── Database ──────────────────────────────────────────────────────────────────
SQLAlchemy==2.0.36
alembic==1.14.1
psycopg2-binary==2.9.10

# ── Async Task Queue ──────────────────────────────────────────────────────────
celery==5.4.0
redis==5.2.1
celery[redis]==5.4.0
flower==2.0.1

# ── Webex Calling API ─────────────────────────────────────────────────────────
wxcadm==4.3.11
requests==2.32.3

# ── Authentication & Security ─────────────────────────────────────────────────
bcrypt==4.2.1
Flask-Bcrypt==1.0.1
PyJWT==2.10.1
pyotp==2.9.0
qrcode[pil]==8.0
cryptography==44.0.0
python-dotenv==1.0.1
itsdangerous==2.2.0

# ── LDAP ──────────────────────────────────────────────────────────────────────
flask-ldap3-login==1.0.3
ldap3==2.9.1

# ── SAML / OIDC SSO ───────────────────────────────────────────────────────────
python3-saml==1.16.0
Authlib==1.4.0

# ── File Handling & Media ─────────────────────────────────────────────────────
Pillow==11.1.0
python-magic==0.4.27
boto3==1.37.14

# ── Scheduling ────────────────────────────────────────────────────────────────
APScheduler==3.11.0
celery[beat]==5.4.0

# ── Email ─────────────────────────────────────────────────────────────────────
Jinja2==3.1.5
premailer==3.10.0

# ── Forms & Validation ───────────────────────────────────────────────────────
email-validator==2.2.0
phonenumbers==8.13.53

# ── ServiceNow Client ────────────────────────────────────────────────────────
pysnow==0.7.2

# ── HTTP & Utils ─────────────────────────────────────────────────────────────
httpx==0.28.1
tenacity==9.0.0
python-dateutil==2.9.0
pytz==2025.1
click==8.1.8

# ── Monitoring & Logging ──────────────────────────────────────────────────────
structlog==25.1.0
colorlog==6.9.0
sentry-sdk[flask]==2.23.1

# ── Certbot (installed system-level but listed for reference) ─────────────────
# certbot  ← installed via OS package manager
# python3-certbot-nginx
